third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/reporting-navigation.https-expected.txt

This is a testharness.js-based test.
PASS parent: (none), target: (none), with empty frame: false
PASS parent: (none), target: (none), with empty frame: true
PASS parent: (none), target: |header(cross-origin-embedder-policy,require-corp), with empty frame: false
PASS parent: (none), target: |header(cross-origin-embedder-policy,require-corp), with empty frame: true
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-embedder-policy,require-corp), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-embedder-policy,require-corp)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-embedder-policy,require-corp), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-embedder-policy,require-corp)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: (none), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: (none), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=" but got (undefined) undefined
PASS parent: (none), target: |header(cross-origin-resource-policy,cross-origin), with empty frame: false
PASS parent: (none), target: |header(cross-origin-resource-policy,cross-origin), with empty frame: true
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-resource-policy,cross-origin), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-resource-policy,cross-origin)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-resource-policy,cross-origin), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-resource-policy,cross-origin)" but got (undefined) undefined
PASS parent: (none), target: |header(cross-origin-embedder-policy,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: false
PASS parent: (none), target: |header(cross-origin-embedder-policy,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: true
PASS parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-embedder-policy,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: false
PASS parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-embedder-policy,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: true
FAIL parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: |header(cross-origin-embedder-policy,require-corp), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-embedder-policy,require-corp)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: |header(cross-origin-embedder-policy,require-corp), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-embedder-policy,require-corp)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: (none), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: (none), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: |header(cross-origin-resource-policy,cross-origin), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-resource-policy,cross-origin)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: |header(cross-origin-resource-policy,cross-origin), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-resource-policy,cross-origin)" but got (undefined) undefined
PASS parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: |header(cross-origin-embedder-policy,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: false
PASS parent: |header(cross-origin-embedder-policy-report-only,require-corp), target: |header(cross-origin-embedder-policy,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: true
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-embedder-policy-report-only,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: false assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-embedder-policy-report-only,require-corp)|header(cross-origin-resource-policy,cross-origin)" but got (undefined) undefined
FAIL parent: |header(cross-origin-embedder-policy,require-corp), target: |header(cross-origin-embedder-policy-report-only,require-corp)|header(cross-origin-resource-policy,cross-origin), with empty frame: true assert_equals: expected (string) "https://www1.web-platform.test:8444/common/blank.html?pipe=|header(cross-origin-embedder-policy-report-only,require-corp)|header(cross-origin-resource-policy,cross-origin)" but got (undefined) undefined
Harness: the test ran to completion.
